Add files via upload
Our website is titled LechEasy.Docs, aiming to simplify requests through Barangay Lecheria's web platform. Its purpose is to establish a comprehensive and user-friendly digital platform for both the residents of Barangay Lecheria and the assigned Barangay officials. Clear policies for acceptable use and behavior are implemented to maintain a secure online environment. The primary focus is on improving communication and streamlining document management processes within the community.